Materials Projects DataBase Dashboard/requirements.txt
dash>=2.5.1
dash-bootstrap-components>=1.0.2
dash-iconify
pandas
plotly>=5.4.0
numpy
Werkzeug==2.0.0
waitress
